+++ client/src/App3.js
@@ -1,0 +1,8 @@
+import React, { useState } from 'react';
+import './App.css';
+
+
+
+export default function App() {
+
+    return(<div>test</div>)}
